requirements.txt
dvc 
mlflow
numpy
pandas
scikit-learn
matplotlib
seaborn
torch
torchvision
nibabel
nilearn
opencv-python
python-dotenv
scikit-image
joblib
boto3
s3fs
